# docs/swagger.yaml
basePath: /
definitions:
...
  model.ApiError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  user.UserCreateRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        description: optional role assignment
        type: string
      username:
        type: string
    type: object
  user.UserResponse:
    properties:
      address_country:
        type: string
      address_locality:
        type: string
      address_postal_code:
        type: string
      address_region:
        type: string
      address_street:
        type: string
      birthdate:
        type: string
      created_at:
        type: string
      email:
        type: string
      failed_login_attempts:
        type: integer
      family_name:
        type: string
      gender:
        type: string
      given_name:
        description: OIDC standard profile claims
        type: string
      id:
        type: string
      is_email_verified:
        type: boolean
      locale:
        type: string
      locked_until:
        type: string
      middle_name:
        type: string
      nickname:
        type: string
      phone_number:
        type: string
      phone_number_verified:
        type: boolean
      picture:
        type: string
      profile:
        type: string
      role:
        type: string
      totp_verified:
        type: boolean
      username:
        type: string
      website:
        type: string
      zoneinfo:
        type: string
    type: object
  user.UserUpdateRequest:
    properties:
      address_country:
        type: string
      address_locality:
        type: string
      address_postal_code:
        type: string
      address_region:
        type: string
      address_street:
        type: string
      birthdate:
        type: string
      email:
        type: string
      family_name:
        type: string
      gender:
        type: string
      given_name:
        description: OIDC standard profile claims
        type: string
      is_email_verified:
        type: boolean
      locale:
        type: string
      middle_name:
        type: string
      nickname:
        type: string
      password:
        type: string
      phone_number:
        type: string
      phone_number_verified:
        type: boolean
      picture:
        type: string
      profile:
        type: string
      role:
        type: string
      totp_verified:
        type: boolean
      username:
        type: string
      website:
        type: string
      zoneinfo:
        type: string
    type: object
host: localhost:9999
info:
  contact: {}
  description: Authentication Service
  title: Autentico OIDC
  version: "1.0"
paths:
...
  /admin/api/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users-admin
    post:
      consumes:
      - application/json
      description: Registers a new user in the system (admin only)
      parameters:
      - description: User creation payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - users-admin
  /admin/api/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users-admin
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - users-admin
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ApiError'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users-admin
  /admin/api/users/{id}/unlock:
    post:
      description: Resets the failed login attempts and clears the lockout time for
        a user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
      security:
      - BearerAuth: []
      summary: Unlock user account
      tags:
      - users-admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"